+++ src/Transformers/IntervalDiscretizer.php
@@ -2,7 +2,6 @@
 
 namespace Rubix\ML\Transformers;
 
-use Tensor\Vector;
 use Rubix\ML\DataType;
 use Rubix\ML\Persistable;
 use Rubix\ML\Helpers\Stats;
@@ -11,6 +10,11 @@ use Rubix\ML\Traits\AutotrackRevisions;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
 use Rubix\ML\Exceptions\InvalidArgumentException;
 use Rubix\ML\Exceptions\RuntimeException;
+
+use function Rubix\ML\linspace;
+use function array_slice;
+use function min;
+use function max;
 
 /**
  * Interval Discretizer
@@ -34,6 +38,13 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
     protected int $bins;
 
     /**
+     * Should the bins be equal width?
+     *
+     * @var bool
+     */
+    protected bool $equiWidth;
+
+    /**
      * The bin intervals of the fitted data.
      *
      * @var array[]|null
@@ -42,9 +53,10 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
 
     /**
      * @param int $bins
+     * @param bool $equiWidth
      * @throws \Rubix\ML\Exceptions\InvalidArgumentException
      */
-    public function __construct(int $bins = 5)
+    public function __construct(int $bins = 5, bool $equiWidth = false)
     {
         if ($bins < 3) {
             throw new InvalidArgumentException('Number of bins must be'
@@ -52,6 +64,7 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
         }
 
         $this->bins = $bins;
+        $this->equiWidth = $equiWidth;
     }
 
     /**
@@ -96,9 +109,11 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
     {
         SamplesAreCompatibleWithTransformer::with($dataset, $this)->check();
 
-        $q = Vector::linspace(0.0, 1.0, 1 + $this->bins)->asArray();
+        if (!$this->equiWidth) {
+            $q = linspace(0.0, 1.0, $this->bins + 1);
 
-        $q = array_slice($q, 1, -1);
+            $q = array_slice($q, 1, -1);
+        }
 
         $this->intervals = [];
 
@@ -106,7 +121,16 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
             if ($type->isContinuous()) {
                 $values = $dataset->feature($column);
 
-                $edges = Stats::quantiles($values, $q);
+                if (isset($q)) {
+                    $edges = Stats::quantiles($values, $q);
+                } else {
+                    $min = min($values);
+                    $max = max($values);
+
+                    $edges = linspace($min, $max, $this->bins + 1);
+
+                    $edges = array_slice($edges, 1, -1);
+                }
 
                 $edges[] = INF;